imgproc: Add FindContoursWithParams function
Until now, the convenient version of `FindContours` was used by gocv. This
PR adds a binding to the full version with the `hierarchy` parameter.

`hierarchy` gives the contours hierarchy, like which contour is the
child of an other.

The current binding of `FindContours` has a lot of code and I tried to not
duplicate it in the `WithParams` version. As a result, now
`FindContours` calls `FindContoursWithParams` but since I didn't write C
code since years now, I'm not sure about how to pass a default
`hierarchy` that would call the convenient version instead of the full
version.

About test, I added a new 1bit image `images/contours.png` in order to
have very simple contours, and so simple hierarchy.